検査実施件数: Rename components/cards/TestedNumberCard.vue to components/index/CardsReference/TestedNumber/Card.vue
