+++ components/index/CardsReference/TestedNumber/Card.vue
